+++ database/migrations/2020_10_01_212551_create_replies_table.php
@@ -19,8 +19,8 @@ class CreateRepliesTable extends Migration
             // Foreing key to calification
             $table->unsignedBigInteger('calification_id');
             $table->foreign('calification_id')->references('id')->on('califications');
-            // Foreing key to products
-            $table->unsignedBigInteger('user_id');
+            // Foreing key to users
+            $table->unsignedInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
         });
     }
